requirements.txt
google-genai
streamlit
faiss-cpu
sentence-transformers
python-dotenv